\name{GENESIS1.2-package}
\alias{GENESIS1.2-package}
\alias{GENESIS1.2}
\docType{package}
\title{
\packageTitle{GENESIS1.2}
}
\description{
\packageDescription{GENESIS1.2}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{GENESIS1.2}
\packageIndices{GENESIS1.2}
GENetic Effect-Size distribution Inference from Summary-level data
}
\author{
\packageAuthor{GENESIS1.2}

Maintainer: \packageMaintainer{GENESIS1.2}
}
\references{
Zhang, Yan, et al. "Estimation of complex effect-size distributions using summary-level statistics from genome-wide association studies across 32 complex traits" Nature Genetics (2018).
}
\keyword{effect size distribution, GWAS, mixture model }
